unify login for ngo and observer
