+++ src/api/VotingIrregularities.Api/Models/AccountViewModels/UserType.cs
@@ -1,0 +1,8 @@
+﻿namespace VotingIrregularities.Api.Models.AccountViewModels
+{
+	public enum UserType
+    {
+        Observer,
+        NgoAdmin
+    }
+}
