+++ src/api/VotingIrregularities.Api/Models/Authentification/AuthenticationResponseModel.cs
@@ -1,0 +1,9 @@
+﻿namespace VotingIrregularities.Api.Models.Authentification
+{
+	public class AuthenticationResponseModel
+	{
+		public string access_token { get; set; }
+		public int expires_in { get; set; }
+
+	}
+}
